fixed product put route

+++ routes/api/product-routes.js
@@ -58,43 +58,46 @@ router.post('/', async (req, res) => {
   }
 });
 
-// update product data
-router.put('/:id', async (req, res) => {
-  try {
-    await Product.update(req.body, {
-      where: {
-        id: req.params.id,
-      },
-    });
-
-    // find all associated tags from ProductTag
-    const productTags = await ProductTag.findAll({ where: { product_id: req.params.id } });
-
-    // get list of current tag_ids
-    const productTagIds = productTags.map(({ tag_id }) => tag_id);
-    // create filtered list of new tag_ids
-    const newProductTags = req.body.tagIds
-      .filter((tag_id) => !productTagIds.includes(tag_id))
-      .map((tag_id) => {
-        return {
-          product_id: req.params.id,
-          tag_id,
-        };
+// update product: how to write as async?
+router.put('/:id', (req, res) => {
+  // update product data
+  Product.update(req.body, {
+    where: {
+      id: req.params.id,
+    },
+  })
+  .then((product) => {
+    if (req.body.tagIds && req.body.tagIds.length) {
+      const productTags = ProductTag.findAll({ where: { product_id: req.params.id } 
       });
-    // figure out which ones to remove
-    const productTagsToRemove = productTags
-      .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-      .map(({ id }) => id);
+      const productTagIds = productTags.map(({ tag_id }) => tag_id);
+      // create filtered list of new tag_ids
+      const newProductTags = req.body.tagIds
+        .filter((tag_id) => !productTagIds.includes(tag_id))
+        .map((tag_id) => {
+          return {
+            product_id: req.params.id,
+            tag_id,
+          };
+        });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
 
-    // run both actions
-    const updatedProductTags = await ProductTag.all([
-      ProductTag.destroy({ where: { id: productTagsToRemove } }),
-      ProductTag.bulkCreate(newProductTags),
-    ]);
-    res.status(200).json(updatedProductTags);
-  } catch (err) {
-    res.status(400).json(err);
-  }
+        // run both actions
+          return Promise.all([
+            ProductTag.destroy({ where: { id: productTagsToRemove } }),
+            ProductTag.bulkCreate(newProductTags),
+          ]);
+        }
+
+        return res.json(product);
+    }) 
+    .catch((err) => {
+      // console.log(err);
+      res.status(400).json(err);
+    });
 });
 
 // delete one product by its `id` value
